Created prefab for columns that triggers adding up the score every time the bird passes through
Based on the official tutorials from the Unity documentation. Link: https://unity3d.com/learn/tutorials/topics/2d-game-creation/adding-column-obstacles?playlist=17093

- Practiced how to trigger a function based on contact of an object with a Collider2D set to be a trigger.
Basically added a set of columns prefab with a collider2D component as a trigger that is able to trigger some code when the bird passes through, and a rigidbody2d component as a kinematic rigidbody that allows to move (assign a constant velocity) the columns prefab via a script.